Add index on EventType.parentId
When including "children" on a Prisma query, Prisma generates a query for child event types whose only condition is `parentId IN (...)`.

Without an index, this is executed as a sequential scan across all entries in the EventType table – which can be slow, and gets slower the more records we have.

From an average of 3 results, this increases eventTypes/getByViewer response times from 134ms to 41ms.